+++ frontend/nextjs-project/src/app/(app)/setting/page.tsx
@@ -1,57 +1,15 @@
-"use client";
+import { Metadata } from "next";
+import SettingClient from "./client";
 
-import SubmitButton from "@/components/ui/SubmitButton";
-import { useAuthStore } from "@/stores/authStore";
-import { Stack } from "@mui/material";
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import OverlayLoader from "@/components/features/loading/OverlayLoader";
-import { downloadRecentPostsJson } from "@/libs/downloadRecentPosts";
+export const metadata: Metadata = {
+  title: "設定 - holo",
+  // ★ 検索エンジンにインデックスさせないための重要な設定
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
 
 export default function SettingPage() {
-  const signOut = useAuthStore((state) => state.signOut);
-  const user = useAuthStore((state) => state.user);
-  const initialized = useAuthStore((state) => state.initialized);
-  const router = useRouter();
-  const [isDownloading, setIsDownloading] = useState<boolean>(false);
-  const handleLogout = async () => {
-    await signOut();
-    alert("ログアウトしました");
-    router.push("/");
-  };
-
-  const handleDownload = async () => {
-    try {
-      setIsDownloading(true);
-      await downloadRecentPostsJson();
-    } catch (e) {
-      alert("ダウンロードに失敗しました");
-      console.error(e);
-    } finally {
-      setIsDownloading(false);
-    }
-  };
-
-  // 万が一URLに直接アクセスされたときの対策
-  useEffect(() => {
-    if (!initialized) {
-      return;
-    }
-    if (user === null) {
-      router.replace("/");
-    }
-  }, [user, initialized, router]);
-
-  if (!initialized || isDownloading) {
-    return <OverlayLoader />;
-  }
-
-  return (
-    <Stack pt={2} spacing={2}>
-      <SubmitButton onClick={handleDownload}>
-        投稿一覧をダウンロード
-      </SubmitButton>
-      <SubmitButton onClick={handleLogout}>ログアウト</SubmitButton>
-    </Stack>
-  );
+  return <SettingClient />;
 }
