+++ frontend/nextjs-project/src/app/login/page.tsx
@@ -1,28 +1,12 @@
-"use client";
+import { Metadata } from "next";
+import LoginClient from "./client";
 
-import LoginView from "@/components/features/auth/LoginView";
-import { useAuthStore } from "@/stores/authStore";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-import OverlayLoader from "@/components/features/loading/OverlayLoader";
+export const metadata: Metadata = {
+  title: "holo - AIカメラと問いで作る新しい散歩地図",
+  description:
+    "散歩×地図×AIカメラで、日常の「どうして？」を見つけにいくサービスです。あなたの「はっけん」をマップに記録・共有しましょう。",
+};
 
 export default function LoginPage() {
-  // ログイン済みならマップにリダイレクト
-  const user = useAuthStore((state) => state.user);
-  const initialized = useAuthStore((state) => state.initialized);
-  const router = useRouter();
-  useEffect(() => {
-    if (!initialized) {
-      return;
-    }
-    if (user !== null) {
-      router.replace("/");
-    }
-  }, [user, initialized, router]);
-
-  if (!initialized || user !== null) {
-    return <OverlayLoader />;
-  }
-
-  return <LoginView />;
+  return <LoginClient />;
 }
